requirements.txt
aniso8601==7.0.0
api==0.0.7
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cognitive-face==1.5.0
Flask==1.0.3
flask-restplus==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
JsonForm==0.0.2
jsonschema==3.0.1
JsonSir==0.0.2
MarkupSafe==1.1.1
nose==1.3.7
Pillow==6.0.0
pyrsistent==0.15.2
Python-EasyConfig==0.1.7
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
Resource==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
